Refactor codebase following Homebridge best practices
Split monolithic 1,234-line index.js into modular structure:
- Created src/platform.js for TccPlatform class
- Created src/accessories/ for accessory classes
- Created src/handlers/ for characteristic handlers
- Created src/helpers/ for utility classes

Benefits:
- Improved code organization and maintainability
- Better separation of concerns
- Easier testing and debugging
- Follows Homebridge verified plugin standards
- Reduced index.js to minimal 18-line entry point

Version bumped to 1.0.0-beta.0 to reflect major architectural improvements.

All tests passing (70/70), 100% coverage maintained, zero linting warnings.
