+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "homebridge-tcc",
-  "version": "0.3.1-beta.10",
+  "version": "1.0.0-beta.0",
   "description": "Honeywell Total Connect Comfort support for Homebridge: https://github.com/thindiyeh/homebridge-tcc",
   "license": "ISC",
   "main": "src/index.js",
